+++ classes/druid.lua
@@ -110,8 +110,8 @@ local function updateSAOs(self)
             updateLeftSAO (self, omenTexture, omenSpellID);
             updateRightSAO(self, omenTexture, omenSpellID);
         else
-            updateLeftSAO (self, '', 0);
-            updateRightSAO(self, '', 0);
+            updateLeftSAO (self, '', nil);
+            updateRightSAO(self, '', nil);
         end
     end
 end
@@ -299,6 +299,12 @@ local function registerClass(self)
     self:RegisterAura("wrath_of_elune", 0, 46833, "shooting_stars", "Top", 1, 255, 255, 255, true, { starfire }); -- PvP season 5-6-7-8
     self:RegisterAura("elunes_wrath", 0, 64823, "shooting_stars", "Top", 1, 255, 255, 255, true, { starfire }); -- PvE tier 8
 
+    -- Fury of Stormrage (Season of Discovery)
+    local furyOfStormrage = 414800;
+    if WOW_PROJECT_ID == WOW_PROJECT_CLASSIC and GetSpellInfo(furyOfStormrage) then
+        self:RegisterAura("fury_of_stormrage", 0, furyOfStormrage, "fury_of_stormrage", "Top", 1, 255, 255, 255, true, { healingTouch });
+    end
+
     -- Healing Trance / Soul Preserver
     self:RegisterAuraSoulPreserver("soul_preserver_druid", 60512); -- 60512 = Druid buff
 
@@ -345,6 +351,9 @@ local function loadOptions(self)
     local wrathOfEluneTalent = wrathOfEluneBuff; -- Not really a talent
     local elunesWrathTalent = elunesWrathBuff; -- Not really a talent
 
+    local furyOfStormrageBuff = 414800;
+    local furyOfStormrageTalent = furyOfStormrageBuff; -- Not really a talent
+
     local predatoryStrikesTalent = 16972;
     local predatoryStrikesBuff = 69369;
 
@@ -359,6 +368,9 @@ local function loadOptions(self)
     self:AddOverlayOption(solarEclipseTalent, solarSpellID, 0, nil, nil, nil, solarSpellID+1000000); -- Spell ID not used by ActivateOverlay like typical overlays
     self:AddOverlayOption(wrathOfEluneTalent, wrathOfEluneBuff);
     self:AddOverlayOption(elunesWrathTalent, elunesWrathBuff);
+    if WOW_PROJECT_ID == WOW_PROJECT_CLASSIC and GetSpellInfo(furyOfStormrageBuff) then
+        self:AddOverlayOption(furyOfStormrageTalent, furyOfStormrageBuff);
+    end
     self:AddOverlayOption(naturesGraceTalent, naturesGraceBuff, 0, nil, naturesGraceVariants);
     self:AddOverlayOption(predatoryStrikesTalent, predatoryStrikesBuff);
     self:AddSoulPreserverOverlayOption(60512); -- 60512 = Druid buff
@@ -367,6 +379,9 @@ local function loadOptions(self)
     self:AddGlowingOption(solarEclipseTalent, wrath, wrath);
     self:AddGlowingOption(wrathOfEluneTalent, wrathOfEluneBuff, starfire);
     self:AddGlowingOption(elunesWrathTalent, elunesWrathBuff, starfire);
+    if WOW_PROJECT_ID == WOW_PROJECT_CLASSIC and GetSpellInfo(furyOfStormrageBuff) then
+        self:AddGlowingOption(furyOfStormrageTalent, furyOfStormrageBuff, healingTouch);
+    end
     if WOW_PROJECT_ID ~= WOW_PROJECT_CLASSIC then -- Must exclude this option specifically for Classic Era, because the talent exists in Era but it has no proc
         self:AddGlowingOption(predatoryStrikesTalent, predatoryStrikesBuff, regrowth);
         self:AddGlowingOption(predatoryStrikesTalent, predatoryStrikesBuff, healingTouch);